requirement.txt
flask
cryptography
mysql-connector-python
qrcode